refactor: move review models to reviews.py
